backend-python/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfplumber
numpy
tqdm
nltk
